Timer WIP
Create separate timer task and build API.
