+++ firmware_mpgl1/application/main.c
@@ -52,7 +52,6 @@ void main(void)
   ClockSetup();
   InterruptSetup();
   SysTickSetup();
-  TimerSetup();
 
   /* Driver initialization */
   MessagingInitialize();
@@ -61,6 +60,7 @@ void main(void)
 
   /* Debug messages through DebugPrintf() are available from here */
 
+  TimerInitialize();  
   SspInitialize();
   TWIInitialize();
 
@@ -88,6 +88,7 @@ void main(void)
     LedUpdate();
     ButtonRunActiveState();
     UartRunActiveState();
+    TimerRunActiveState(); 
     SspRunActiveState();
     TWIRunActiveState();
     MessagingRunActiveState();
