+++ firmware_mpgl1/bsp/mpgl1-ehdw-03.c
@@ -272,116 +272,6 @@ void GpioSetup(void)
 } /* end GpioSetup() */
 
 
-/*----------------------------------------------------------------------------------------------------------------------
-Function: TimerSetup
-
-Description
-Configures Timer Counter (TC) peripherals.
-
-Requires:
-  - Peripheral resources not used for any other function.
-
-Promises:
-  - Timer1 channel is fully configured (running state and interrupt state
-    determined by 
-*/
-void TimerSetup(void)
-{
-  /* Load the block configuration register */
-  AT91C_BASE_TCB0->TCB_BMR = TCB_BMR_INIT;
-
-  /* Channel 0 settings not configured at this time */
-  
-  /* Load Channel 1 settings */
-  AT91C_BASE_TC1->TC_CMR = TC1_CMR_INIT;
-  AT91C_BASE_TC1->TC_RC  = TC1_RC_INIT;
-  AT91C_BASE_TC1->TC_IER = TC1_IER_INIT;
-  AT91C_BASE_TC1->TC_IDR = TC1_IDR_INIT;
-  
-  /* Activate the timer clock */
-  AT91C_BASE_TC1->TC_CCR = TC1_CCR_INIT;
-
-  /* Channel 2 settings not configured at this time */
-
-  /* Enable required interrupts */
-  NVIC_ClearPendingIRQ(IRQn_TC1);
-  NVIC_EnableIRQ(IRQn_TC1);
-}
-
-/* end Timer0Setup() */
-
-#if 0
-/*----------------------------------------------------------------------------------------------------------------------
-Function: Timer0Init
-
-Description
-Sets the designated Timer period, on/off status, and interrupt status.
-
-Requires:
-  - eTimerChannel_ is the timer to be set in TC0
-  - psTimerSettings_ points to TimerSetupType holding the period, timer state, and timer interrupt state
-  - Peripheral resources not used for any other function.
-
-Promises:
-  - Specified timer is set according to psTimerSettings_
-*/
-void TimerInit(TimerType eTimer_, TimerSetupType* psTimerSettings_)
-{
-  AT91C_BASE_TC0->TC_CMR[eChannel_] = 
-
-
-} /* end TimerInit */
-#endif
-
-/*----------------------------------------------------------------------------------------------------------------------
-Function: TimerStart
-
-Description
-Starts the designated Timer.
-
-Requires:
-  - eTimer_ is the timer to start
-
-Promises:
-  - Specified channel on Timer 0 is set to run; if already running it remains running
-  - Does NOT reset the timer value
-*/
-void TimerStart(TimerChannelType eTimerChannel_)
-{
-  u32 u32TimerBaseAddress = (u32)AT91C_BASE_TC0;
-  u32TimerBaseAddress += (u32)eTimerChannel_;
-
-  /* Ensure clock is enabled */
-  (AT91_CAST(AT91PS_TC)u32TimerBaseAddress)->TC_CCR = (AT91C_TC_CLKEN | AT91C_TC_SWTRG);
-  
-} /* end TimerStart() */
-
-
-/*----------------------------------------------------------------------------------------------------------------------
-Function: TimerStop
-
-Description
-Stops the designated Timer.
-
-Requires:
-  - eTimer_ is the timer to stop
-
-Promises:
-  - Specified timer is stopped; if already stopped it remains stopped
-  - Does NOT reset the timer value
-*/
-void TimerStop(TimerChannelType eTimerChannel_)
-{
-  u32 u32TimerBaseAddress = (u32)AT91C_BASE_TC0;
-  u32TimerBaseAddress += (u32)eTimerChannel_;
-  
-  /* Ensure clock is disabled */
-  (AT91_CAST(AT91PS_TC)u32TimerBaseAddress)->TC_CCR = AT91C_TC_CLKDIS;
-  
-} /* end TimerStop */
-
-
-
 /*----------------------------------------------------------------------------
 Function: PWMSetupAudio
