+++ firmware_mpgl1/iar_7_20_1/mpgl1-efmw-01.ewp
@@ -1982,6 +1982,9 @@
         <name>$PROJ_DIR$\..\drivers\sdcard.h</name>
       </file>
       <file>
+        <name>$PROJ_DIR$\..\..\firmware_mpg_common\drivers\timer.h</name>
+      </file>
+      <file>
         <name>$PROJ_DIR$\..\..\firmware_mpg_common\drivers\utilities.h</name>
       </file>
     </group>
@@ -2022,6 +2025,9 @@
       </file>
       <file>
         <name>$PROJ_DIR$\..\drivers\sdcard.c</name>
+      </file>
+      <file>
+        <name>$PROJ_DIR$\..\..\firmware_mpg_common\drivers\timer.c</name>
       </file>
       <file>
         <name>$PROJ_DIR$\..\..\firmware_mpg_common\drivers\utilities.c</name>
