+++ firmware_mpgl1/iar_7_20_1/settings/mpgl1-efmw-01.wspos
@@ -1,6 +1,6 @@
 [MainWindow]
 <<<<<<< HEAD
-WindowPlacement=_ 465 5 1924 1013 1
+WindowPlacement=_ 83 226 793 504 3
 =======
 WindowPlacement=_ 118 95 2003 1316 1
 >>>>>>> origin/master
